Modified Base & Enemies
Modified GameView so that multiple enemies of each type can spawn.
Modified Base so that it can take damage and be “killed”.
Adjusted damage of each enemy class
Modified Enemy class so that enemies will spawn randomly from -500 to 0
so that they no longer stack on top of each other.